lenta_app/requirements.txt
Django~=5.0.6
pathlib~=1.0.1
Pillow
